Step2 코드 작성
