+++ CodeStarterCamp_Week4/Activity.swift
@@ -1,0 +1,38 @@
+//
+//  Activity.swift
+//  CodeStarterCamp_Week4
+//
+//  Created by Dawn on 2023/03/20.
+//
+
+import Foundation
+
+struct Activity {
+    let name: String
+    let action: (BodyCondition) -> Void
+}
+
+let sitUp: Activity = Activity(name: "윗몸일으키기", action: { bodyCondition in
+    print("<<\(sitUp.name)을(를) 시작합니다>>")
+    bodyCondition.upperBodyStrength += Int.random(in: 10...20)
+    bodyCondition.fatigue += Int.random(in: 10...20)
+})
+
+let squat: Activity = Activity(name: "스쿼트", action: { bodyCondition in
+    print("<<\(squat.name)을(를) 시작합니다>>")
+    bodyCondition.lowerBodyStrength += Int.random(in: 20...30)
+    bodyCondition.fatigue += Int.random(in: 10...20)
+})
+
+let longRun: Activity = Activity(name: "오래달리기", action: { bodyCondition in
+    print("<<\(longRun.name)을(를) 시작합니다>>")
+    bodyCondition.muscularEndurance += Int.random(in: 20...30)
+    bodyCondition.upperBodyStrength += Int.random(in: 5...10)
+    bodyCondition.lowerBodyStrength += Int.random(in: 5...10)
+    bodyCondition.fatigue += Int.random(in: 20...30)
+})
+
+let dynamicBreak: Activity = Activity(name: "동적휴식", action: { bodyCondition in
+    print("<<\(dynamicBreak.name)을(를) 시작합니다>>")
+    bodyCondition.fatigue -= Int.random(in: 5...10)
+})
