+++ CodeStarterCamp_Week4/RoutineError.swift
@@ -1,0 +1,12 @@
+//
+//  RoutineError.swift
+//  CodeStarterCamp_Week4
+//
+//  Created by Dawn on 2023/03/20.
+//
+
+import Foundation
+
+enum RoutineError: Error {
+    case exceededFatigue, invalidInput, exceededTimes
+}
